pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["llmSHAP*"]

[project]
name = "llmSHAP"
version = "0.1.0"
description = "llmSHAP"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Filip Naudot" },
]
license = { text = "MIT" }

dependencies = [
  "scikit-learn",
]

[project.optional-dependencies]
openai     = ["openai", "python-dotenv"]
embeddings = ["sentence-transformers"]
dev        = ["pytest", "matplotlib", "sphinx", "myst-parser", "sphinx-book-theme", "sphinx-design"]
full       = ["llmSHAP[openai]", "llmSHAP[embeddings]", "llmSHAP[dev]"]
notebook   = ["ipywidgets"]